requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.41
razorpay==1.4.0
python-dotenv==0.21.0
requests==2.31.0
beautifulsoup4==4.12.2
uuid==1.30
